move FMI code in cpp files
